requirements.txt
pandas
numpy
openai
faiss-cpu
